requirements.txt
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.22.0
cffi==1.10.0
click==6.7
cryptography==1.8.1
decorator==4.0.11
Fabric3==1.13.1.post1
flake8==3.3.0
Flask==0.12.1
idna==2.5
ipython==5.3.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mccabe==0.6.1
packaging==16.8
paramiko==2.1.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pyasn1==0.2.3
pycodestyle==2.3.1
pycparser==2.17
pyflakes==1.5.0
Pygments==2.2.0
pyparsing==2.2.0
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.1
